+++ cmd/wit-bindgen-go/cmd/wit/wit.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/urfave/cli/v3"
 	"github.com/bytecodealliance/wasm-tools-go/internal/witcli"
+	"github.com/urfave/cli/v3"
 )
 
 // Command is the CLI command for wit.
@@ -16,7 +16,11 @@ var Command = &cli.Command{
 }
 
 func action(ctx context.Context, cmd *cli.Command) error {
-	res, err := witcli.LoadOne(cmd.Bool("force-wit"), cmd.Args().Slice()...)
+	path, err := witcli.LoadPath(cmd.Args().Slice()...)
+	if err != nil {
+		return err
+	}
+	res, err := witcli.LoadWIT(ctx, cmd.Bool("force-wit"), path)
 	if err != nil {
 		return err
 	}