+++ internal/witcli/witcli.go
@@ -1,19 +1,41 @@
 package witcli
 
 import (
+	"context"
 	"fmt"
+	"os"
 	"strings"
 
+	"github.com/bytecodealliance/wasm-tools-go/internal/oci"
 	"github.com/bytecodealliance/wasm-tools-go/wit"
 )
 
-// LoadOne loads a single [wit.Resolve].
-// An error is returned if len(paths) > 1.
-// If paths is empty, or paths[0] == "" or "-", then it reads from stdin.
+// LoadWIT loads a single [wit.Resolve].
+// If path is a OCI path, it pulls from the OCI registry and load WIT
+// from the buffer.
+// If path == "" or "-", then it reads from stdin.
 // If the resolved path doesn’t end in ".json", it will attempt to load
 // WIT indirectly by processing the input through wasm-tools.
 // If forceWIT is true, it will always process input through wasm-tools.
-func LoadOne(forceWIT bool, paths ...string) (*wit.Resolve, error) {
+func LoadWIT(ctx context.Context, forceWIT bool, path string) (*wit.Resolve, error) {
+	if oci.IsOCIPath(path) {
+		fmt.Fprintf(os.Stderr, "Fetching OCI artifact %s\n", path)
+		if bytes, err := oci.PullWIT(ctx, path); err != nil {
+			return nil, err
+		} else {
+			return wit.ParseWIT(bytes)
+		}
+	}
+	if forceWIT || !strings.HasSuffix(path, ".json") {
+		return wit.LoadWIT(path)
+	}
+	return wit.LoadJSON(path)
+}
+
+// LoadPath parses paths and returns the first path.
+// If paths is empty, returns "-".
+// If paths has more than one element, returns an error.
+func LoadPath(paths ...string) (string, error) {
 	var path string
 	switch len(paths) {
 	case 0:
@@ -21,10 +43,7 @@ func LoadOne(forceWIT bool, paths ...string) (*wit.Resolve, error) {
 	case 1:
 		path = paths[0]
 	default:
-		return nil, fmt.Errorf("found %d path arguments, expecting 0 or 1", len(paths))
+		return "", fmt.Errorf("found %d path arguments, expecting 0 or 1", len(paths))
 	}
-	if forceWIT || !strings.HasSuffix(path, ".json") {
-		return wit.LoadWIT(path)
-	}
-	return wit.LoadJSON(path)
+	return path, nil
 }