+++ wit/load.go
@@ -2,7 +2,9 @@ package wit
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
 )
@@ -12,8 +14,9 @@ import (
 //
 // [WIT]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/WIT.md
 func LoadJSON(path string) (*Resolve, error) {
-	if path == "" || path == "-" {
-		return DecodeJSON(os.Stdin)
+	r := reader(path)
+	if r != nil {
+		return DecodeJSON(r)
 	}
 	f, err := os.Open(path)
 	if err != nil {
@@ -30,28 +33,60 @@ func LoadJSON(path string) (*Resolve, error) {
 // [WIT]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/WIT.md
 // [wasm-tools]: https://crates.io/crates/wasm-tools
 func LoadWIT(path string) (*Resolve, error) {
+	r := reader(path)
+	return loadWIT(path, r)
+}
+
+// ParseWIT parses [WIT] data from a buffer by processing it through [wasm-tools].
+// This will fail if wasm-tools is not in $PATH.
+//
+// [WIT]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/WIT.md
+// [wasm-tools]: https://crates.io/crates/wasm-tools
+func ParseWIT(buffer []byte) (*Resolve, error) {
+	r := bytes.NewReader(buffer)
+	return loadWIT("", r)
+}
+
+// loadWIT loads WIT data from path or reader by processing it through wasm-tools.
+// It accepts either a path or an io.Reader as input, but not both.
+// If the path is not "" and "-", it will be used as the input file.
+// Otherwise, the reader will be used as the input.
+func loadWIT(path string, reader io.Reader) (*Resolve, error) {
+	if path != "" && reader != nil {
+		return nil, errors.New("cannot set both path and reader; provide only one")
+	}
+
 	wasmTools, err := exec.LookPath("wasm-tools")
 	if err != nil {
 		return nil, err
 	}
 
-	var stdout bytes.Buffer
-	var stderr bytes.Buffer
+	var stdout, stderr bytes.Buffer
 
-	cmd := exec.Command(wasmTools, "component", "wit", "-j", "--all-features")
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	if path == "" || path == "-" {
-		cmd.Stdin = os.Stdin
-	} else {
-		cmd.Args = append(cmd.Args, path)
+	cmdArgs := []string{"component", "wit", "-j", "--all-features"}
+	if path != "" && path != "-" {
+		cmdArgs = append(cmdArgs, path)
 	}
 
-	err = cmd.Run()
-	if err != nil {
+	cmd := exec.Command(wasmTools, cmdArgs...)
+	cmd.Stdout = &stdout
+	cmd.Stderr = &stderr
+
+	if reader != nil {
+		cmd.Stdin = reader
+	}
+
+	if err := cmd.Run(); err != nil {
 		fmt.Fprint(os.Stderr, stderr.String())
 		return nil, err
 	}
 
 	return DecodeJSON(&stdout)
+}
+
+func reader(path string) io.ReadCloser {
+	if path == "" || path == "-" {
+		return os.Stdin
+	}
+	return nil
 }